Move aura logic check to base Card class
Auras are not just for minions.
Well, that's a bit theoretical anyway. We're doing this because we
want Enrages to be able to have Auras, but there is no pre-gvg card
that actually has an aura other than minions.
In GVG, Cogmaster's Wrench (Weapon) does have an Aura which makes a
lot of sense.

That said, Spells cannot have auras and we should split those out
into a class that has less "I'm always on the board" logic.

Note that because of this, we have to check that the source of an
aura is a minion before checking for adjacentBuff as the aura can
now come from a non-field entity.