Allow for callables in getIntProperty()
Some tags can now implicitly be functions, as long as they are called
through getIntProperty().
The function will take the current value as input, and will output the
result of a calculation. This matters for Gahz'rilla and Blessed Champion
where the attack up to the point of the buff is doubled.

Thanks, Xinhuan